util: move UTF-8 decoding to its own module
This moves UTF-8 decoding from an internal helper in
`jiff::util::escape` to a crate-wide helper in `jiff::util::utf8`.

This makes it easier for other code to use this. I may end up using it
for #111, for example.
